refactor: `MenuGroupService` 분리함으로써 책임 분리
