+++ src/main/java/kitchenpos/menu/application/MenuGroupService.java
@@ -1,0 +1,39 @@
+package kitchenpos.menu.application;
+
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroupRepository;
+import kitchenpos.menu.dto.MenuGroupRequest;
+import kitchenpos.menu.dto.MenuGroupResponse;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+import static java.util.stream.Collectors.toList;
+
+@Service
+public class MenuGroupService {
+    private final MenuGroupRepository menuGroupRepository;
+
+    public MenuGroupService(final MenuGroupRepository menuGroupRepository) {
+        this.menuGroupRepository = menuGroupRepository;
+    }
+
+    @Transactional
+    public MenuGroupResponse create(final MenuGroupRequest menuGroupRequest) {
+        MenuGroup menuGroup = MenuGroup.of(menuGroupRequest.getName());
+        menuGroup = menuGroupRepository.save(menuGroup);
+        return MenuGroupResponse.of(menuGroup);
+    }
+
+    public List<MenuGroupResponse> list() {
+        return menuGroupRepository.findAll().stream()
+                .map(MenuGroupResponse::of)
+                .collect(toList());
+    }
+
+    public MenuGroup findById(Long id) {
+        return menuGroupRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("등록되지 않은 메뉴가 있습니다."));
+    }
+}
